Closed #16 + more
Server is doing exactly the same damage like the client is showing
+ Abyss behavior from serphas :wink:

+++ wServer/logic/behaviors/Shoot.cs
@@ -95,7 +95,7 @@ namespace wServer.logic.behaviors
                     for (int i = 0; i < count; i++)
                     {
                         Projectile prj = host.CreateProjectile(
-                            desc, (short)host.ObjectType, dmg, time.tickTimes,
+                            desc, host.ObjectType, dmg, time.tickTimes,
                             prjPos, (float) (startAngle + shootAngle*i));
                         host.Owner.EnterWorld(prj);
                         if (i == 0)
